+++ src/main/scala/xitrum/metrics/MetricsAction.scala
@@ -140,7 +140,7 @@ class XitrumMetricsViewer extends FutureAction with MetricsViewer {
       <head>
         {xitrumCss}
         {jsDefaults}
-        <script type="text/javascript" src={webJarsUrl("d3js/3.5.17", "d3.js", "d3.min.js")}></script>
+        <script type="text/javascript" src={webJarsUrl("d3/4.1.1", "d3.js", "d3.min.js")}></script>
         <script type="text/javascript" src={webJarsUrl(s"xitrum/${xitrum.version}/metrics.js")}></script>
         <title>Xitrum Default Metrics Viewer</title>
       </head>
@@ -202,7 +202,7 @@ class XitrumMetricsViewer extends FutureAction with MetricsViewer {
       <head>
         {xitrumCss}
         {jsDefaults}
-        <script type="text/javascript" src={webJarsUrl("d3js/3.5.17", "d3.js", "d3.min.js")}></script>
+        <script type="text/javascript" src={webJarsUrl("d3/4.1.1", "d3.js", "d3.min.js")}></script>
         <script type="text/javascript" src={webJarsUrl(s"xitrum/${xitrum.version}/metrics.js")}></script>
         <title>Xitrum Default Metrics Viewer</title>
       </head>